+++ src/main/java/com/github/games647/lagmonitor/TpsHistoryTask.java
@@ -1,19 +1,26 @@
 package com.github.games647.lagmonitor;
 
+import com.google.common.collect.Lists;
+
+import java.util.List;
+
 public class TpsHistoryTask implements Runnable {
 
-    private float lastTicks = 20.0F;
+    private final List<Float> lastSeconds = Lists.newArrayListWithExpectedSize(60);
 
     //the last time we updated the ticks
-    private long lastCheck;
+    private long lastCheck = System.nanoTime();
 
-    /**
-     * Get the ticks count of the last check. 20 Ticks should pass per second
-     *
-     * @return the ticks count of the last check
-     */
-    public float getLastTicks() {
-        return lastTicks;
+    public float getLastSample() {
+        if (lastSeconds.isEmpty()) {
+            return 20F;
+        }
+
+        return lastSeconds.get(lastSeconds.size() - 1);
+    }
+
+    public List<Float> getLastSeconds() {
+        return lastSeconds;
     }
 
     @Override
@@ -25,10 +32,13 @@ public class TpsHistoryTask implements Runnable {
         lastCheck = currentTime;
 
         //how many ticks passed since the last check * 1000 to convert to seconds
-        float tps = 3 * 20 * 1000.0F / (timeSpent / (1000 * 1000));
-        if (tps >= 0.0F && tps < 20.0F) {
+        float tps = 1 * 20 * 1000.0F / (timeSpent / (1000 * 1000));
+        if (tps >= 0.0F && tps < 25.0F) {
             //Prevent all invalid values
-            lastTicks = tps;
+            lastSeconds.add(tps);
+            if (lastSeconds.size() >= 61) {
+                lastSeconds.remove(0);
+            }
         }
     }
 }